fix: Consolidate registry creation and prevent duplicate logs
- Move create_registry to shared handlers module
- Reuse registry instance between workers in worker process
- Remove redundant registry creation code in main.py and worker.py

Closes #203, Closes #204
